requirements.txt
streamlit>=1.31.0
google-generativeai>=0.8.5
pdfplumber>=0.11.7
python-docx>=1.1.0
plotly>=5.15.0
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.0.0
python-dotenv>=1.0.0
nltk>=3.8.0
scikit-learn>=1.3.0
wordcloud>=1.9.2
textstat>=0.7.3
pyyaml>=6.0.1
jinja2>=3.1.2
reportlab>=4.0.4
